style(all):字段名称修改

+++ services/service-auth/src/main/java/qtx/cloud/auth/controller/SysRoleController.java
@@ -73,7 +73,7 @@ public class SysRoleController {
 
     @ApiOperation("获取权限配置")
     @PostMapping("/getMenuByRoleId")
-    public Result<List<SysRoleMenuVo>> getMenuByRole(@RequestParam Long id) {
+    public Result<List<SysRoleMenuVo>> getMenuByRole(@RequestParam Integer id) {
         return Result.success(service.getMenuByRole(id));
     }
